+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- replaced MD5 hashing of `fsm.StateMachine` with blake2b.
+
 ## [0.1.0a8] - 2026-03-04
 
 Just a quick bugfix release ...